requirements.txt
ccxt
numpy
scikit-learn
finta
python-dotenv
scikit-optimize
pandas
